dtb: gpio: drive WLAN_EN and BT_EN to high to power wireless module
After Linux boots up, WLAN_EN and BT_EN line may not be in correct states to gaurantee that the wireless chip is in power off state.
This patch drives those lines to high to provide a low voltage
level to WLAN_REG_ON and BT_REG_ON pin of the wireless chip to make the
chip in power off state after Linux boots up. When wireless driver is
started, the driver will control those lines to power on/off the
wireless chip